# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Syspro APS Scheduler API
  version: "1.0.0"
  description: |
    Finite-capacity production scheduler that reads master data from SYSPRO
    and writes approved schedules back via the APS compatibility layer.
  contact:
    name: APS Development Team
servers:
  - url: http://localhost:3000/api
    description: Local development / production (same port)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Job:
      type: object
      properties:
        jobId:       { type: string }
        itemCode:    { type: string }
        description: { type: string }
        quantity:    { type: number }
        priority:    { type: integer }
        status:      { type: string }
        dueDate:     { type: string, format: date-time }
        releaseDate: { type: string, format: date-time }
        operations:  { type: array, items: { $ref: '#/components/schemas/Operation' } }
    Operation:
      type: object
      properties:
        opId:          { type: string }
        sequence:      { type: integer }
        description:   { type: string }
        workcentreId:  { type: string }
        setupTime:     { type: number, description: Minutes }
        runTime:       { type: number, description: Minutes per unit }
        quantity:      { type: number }
        status:        { type: string }

    # ── Resources ─────────────────────────────────────────────────────────────
...
    Scenario:
      type: object
      properties:
        scenarioId:     { type: string, format: uuid }
        baseScheduleId: { type: string, format: uuid }
        name:           { type: string }
        description:    { type: string }
        status:         { type: string, enum: [draft, promoted, archived] }
        createdBy:      { type: string }
        createdAt:      { type: string, format: date-time }
        promotedAt:     { type: string, format: date-time, nullable: true }

    # ── Errors ────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        error:   { type: string }
        code:    { type: string }
        details: { type: object }
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Insufficient role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ServiceUnavailable:
      description: Database not connected
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /jobs:
    get:
      summary: List open production jobs from SYSPRO
      tags: [Jobs]
      responses:
        "200":
          description: Array of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs: { type: array, items: { $ref: '#/components/schemas/Job' } }
        "503":
          $ref: '#/components/responses/ServiceUnavailable'
...
  /audit:
    get:
      summary: Retrieve audit log entries (authenticated)
      tags: [Audit]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Audit entries
        "401":
          $ref: '#/components/responses/Unauthorized'

  # SSE — real-time replan signals (#63)
...
  /scenarios:
    get:
      summary: List all scenarios (optionally filtered by base schedule)
      tags: [Scenarios]
      parameters:
        - in: query
          name: baseScheduleId
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Scenario list
          content:
            application/json:
              schema:
                type: object
                properties:
                  scenarios:
                    type: array
                    items: { $ref: '#/components/schemas/Scenario' }
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Clone a saved schedule into a new named scenario (Approver role)
      tags: [Scenarios]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [baseScheduleId, name]
              properties:
                baseScheduleId: { type: string, format: uuid }
                name:           { type: string, maxLength: 120 }
                description:    { type: string }
      responses:
        "201":
          description: Scenario created
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /scenarios/{scenarioId}/promote:
    post:
      summary: Promote scenario to live schedule (Approver role)
      tags: [Scenarios]
      parameters:
        - in: path
          name: scenarioId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Scenario promoted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
  /scenarios/{scenarioId}:
    delete:
      summary: Delete a scenario (Approver role)
      tags: [Scenarios]
      parameters:
        - in: path
          name: scenarioId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Scenario deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'